Fixing Order of Events
Now the story is told with all cards being used (we can change the story later) and it brings users to the map, not a specific marker. (It used to be Fisk) Harvesting resources now brings you to the result of your harvest instead of Weber Arch. Weber Arch now takes you to harvesting resources when you click "harvest resources" instead of seeing the resources earned, as it did previously.